requirements.txt
flask==2.0.1
gunicorn==20.1.0
requests==2.31.0
bleach==6.0.0
pywebpush==1.14.0
psycopg2-binary==2.9.6